1.1.2-rc.1 released

+++ hyperswitch-sdk-android-lite/src/main/java/io/hyperswitch/payments/expresscheckoutlauncher/ExpressCheckoutPaymentMethodLauncher.kt
@@ -1,7 +1,7 @@
 package io.hyperswitch.payments.expresscheckoutlauncher
 
 import android.os.Parcelable
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 import kotlinx.parcelize.Parcelize
 
 class ExpressCheckoutPaymentMethodLauncher {
@@ -51,7 +51,7 @@ class ExpressCheckoutPaymentMethodLauncher {
      * See the corresponding [Result.Failed.error] message for more details.
      */
     @Target(AnnotationTarget.PROPERTY, AnnotationTarget.VALUE_PARAMETER, AnnotationTarget.TYPE)
-    @IntDef(INTERNAL_ERROR.toLong(), DEVELOPER_ERROR.toLong(), NETWORK_ERROR.toLong())
+    @IntDef(INTERNAL_ERROR, DEVELOPER_ERROR, NETWORK_ERROR)
     annotation class ErrorCode
 
     companion object {
